+++ src/Description/DataTransferObjects/LanguageData.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Grixu\SociusClient\Description\DataTransferObjects;
-
-class LanguageData extends \Grixu\SociusModels\Description\DataTransferObjects\LanguageData
-{
-    public int $id;
-}