+++ src/Operator/DataTransferObjects/BranchData.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Grixu\SociusClient\Operator\DataTransferObjects;
-
-class BranchData extends \Grixu\SociusModels\Operator\DataTransferObjects\BranchData
-{
-    public int $id;
-    public ?OperatorDataCollection $operators;
-}