+++ src/Product/DataTransferObjects/ProductTypeData.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Grixu\SociusClient\Product\DataTransferObjects;
-
-class ProductTypeData  extends \Grixu\SociusModels\Product\DataTransferObjects\ProductTypeData
-{
-    public int $id;
-    public ?ProductDataCollection $products;
-}