+++ src/SociusClient.php
@@ -2,69 +2,53 @@
 
 namespace Grixu\SociusClient;
 
+use Grixu\ApiClient\JsonApiFetcher;
+
 /**
- * Class Socius
- * @package App\Socius
+ * @method product
+ * @method product_type
+ * @method brand
+ * @method category
+ * @method operator
+ * @method operator_role
+ * @method branch
+ * @method language
+ * @method description
+ * @method warehouse
+ * @method stock
  */
 class SociusClient
 {
-    public function product()
+    public function __call(string $name, array $parameters): JsonApiFetcher
     {
-        return new SociusQuery(SociusDomainEnum::PRODUCT());
+        $this->validateConfig($name);
+
+        return new JsonApiFetcher(
+            JsonApiConfigFactory::makeConfig($name),
+            config("socius-client.{$name}.url")
+        );
     }
 
-    public function productType()
+    protected function validateConfig($name): void
     {
-        return new SociusQuery(SociusDomainEnum::PRODUCT_TYPE());
-    }
+        if (empty(config("socius-client.{$name}"))) {
+            throw new \Exception("Module {name} is not configured.");
+        }
 
-    public function brand()
-    {
-        return new SociusQuery(SociusDomainEnum::BRAND());
-    }
+        if (empty(config("socius-client.{$name}.url"))) {
+            throw new \Exception("Module {$name}: Url is not configured.");
+        }
 
-    public function category()
-    {
-        return new SociusQuery(SociusDomainEnum::CATEGORY());
-    }
+        if (!is_array(config("socius-client.{$name}.filters"))) {
+            throw new \Exception("Module {$name}: Filters is not configured.");
+        }
 
-    public function description()
-    {
-        return new SociusQuery(SociusDomainEnum::DESCRIPTION());
-    }
+        if (!is_array(config("socius-client.{$name}.sorts"))) {
+            throw new \Exception("Module {$name}: Sorts is not configured.");
+        }
 
-    public function language()
-    {
-        return new SociusQuery(SociusDomainEnum::LANGUAGE());
-    }
-
-    public function warehouse()
-    {
-        return new SociusQuery(SociusDomainEnum::WAREHOUSE());
-    }
-
-    public function stock()
-    {
-        return new SociusQuery(SociusDomainEnum::STOCK());
-    }
-
-    public function operator()
-    {
-        return new SociusQuery(SociusDomainEnum::OPERATOR());
-    }
-
-    public function operatorRole()
-    {
-        return new SociusQuery(SociusDomainEnum::OPERATOR_ROLE());
-    }
-
-    public function branch()
-    {
-        return new SociusQuery(SociusDomainEnum::BRANCH());
-    }
-
-    public function customer()
-    {
-        return new SociusQuery(SociusDomainEnum::CUSTOMER());
+        if (!is_array(config("socius-client.{$name}.includes"))) {
+            throw new \Exception("Module {$name}: Includes is not configured.");
+        }
     }
 }