+++ src/SociusClientFacade.php
@@ -4,10 +4,6 @@ namespace Grixu\SociusClient;
 
 use Illuminate\Support\Facades\Facade;
 
-/**
- * Class SociusFacade
- * @package App\Socius
- */
 class SociusClientFacade extends Facade
 {
     protected static function getFacadeAccessor()